# example/server/docs/swagger.yaml
definitions:
  ensweb.ErrMessage:
    properties:
      Message:
        type: string
    type: object
  main.Request:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  main.Response:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/home:
    get:
      consumes:
      - application/json
      description: Login session in the dashbaord
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ensweb.ErrMessage'
      summary: Login Session
      tags:
      - accounts
  /api/login:
    post:
      consumes:
      - application/json
      description: login in the dashboard
      parameters:
      - description: Login Credential
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/main.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ensweb.ErrMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ensweb.ErrMessage'
      summary: Login into account
      tags:
      - accounts
  /api/logout:
    post:
      consumes:
      - application/json
      description: Logout from the session
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "507":
          description: Insufficient Storage
          schema:
            $ref: '#/definitions/ensweb.ErrMessage'
      summary: Logout from the session
      tags:
      - accounts
  /api/register:
    post:
      consumes:
      - application/json
      description: Register new account on the dashboard
      parameters:
      - description: User email & Password
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/main.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ensweb.ErrMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ensweb.ErrMessage'
      summary: Register new user account
      tags:
      - accounts
swagger: "2.0"